backend/requirements.txt
distlib==0.3.8
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
pre-commit==3.8.0
PyJWT==2.9.0
PyYAML==6.0.2
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9

